Code organizing  - move common ui objects into domain/common  - move tests to test folder
